requirements.txt
# Function dependencies, for example:
# package>=version
pyaml-env==1.2.1
langchain==0.0.128
openai==0.27.2